Apply requested changes
