+++ packages/expo-modules-core/android/CMakeLists.txt
@@ -164,4 +164,3 @@ else ()
             android
     )
 endif ()
-
