+++ packages/expo-modules-core/android/src/main/cpp/JNIInjector.cpp
@@ -5,7 +5,7 @@
 #include "JavaScriptValue.h"
 #include "JavaScriptObject.h"
 #include "JavaScriptTypedArray.h"
-#include "CachedReferencesRegistry.h"
+#include "JavaReferencesCache.h"
 
 #include <jni.h>
 #include <fbjni/fbjni.h>
@@ -14,7 +14,7 @@
 JNIEXPORT jint JNICALL JNI_OnLoad(JavaVM *vm, void *) {
   return facebook::jni::initialize(vm, [] {
     // Loads references to often use Java classes
-    expo::JavaCachedReferencesRegistry::instance()->loadJClasses(jni::Environment::current());
+    expo::JavaReferencesCache::instance()->loadJClasses(jni::Environment::current());
 
     expo::JSIInteropModuleRegistry::registerNatives();
     expo::JavaScriptModuleObject::registerNatives();
